Added Login, and course visibility
Added login visibility for users to know when they were logged in, and
when they were not on login screen

Added course visibility to know when a course they were adding was
already in the system. Additionally to know when a course they added was
actually being put into the system.

Added butttons for tutor signup and to view the available tutors for a
course.